pyproject.toml
[tool.poetry]
name = "projecthabits"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-cors-headers = "^4.7.0"
celery = "^5.5.3"
dotenv = "^0.9.9"
djangorestframework = "3.16.0"
djangorestframework-simplejwt = "5.5.0"
django-filter = "25.1"
drf-yasg = "1.21.10"
django-celery-beat = "2.8.1"
django-extensions = "^4.1"
python-telegram-bot = "13.7"
urllib3 = "<2.0.0"
pillow = "^11.2.1"
redis = "6.2.0"
psycopg2-binary = "^2.9.10"
flake8 = "^7.3.0"
black = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
